Add files via upload
This is just a DEMO extension.

Although this is not the final product, it already contains deep links that we are tracking, and you can already use it. While you cannot collect tokens yet, this version already contains our partners and shows the future cashback rate, so you can support us by clicking on the brands and making a purchase.